requirements.txt
# Core monitoring dependencies
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.21.0

# Fast compression library
zstandard>=0.19.0

# Configuration parsing
PyYAML>=6.0
